Database/streamlit/stockpilot/pyproject.toml
[project]
name = "stockpilot"
version = "0.1.0"
description = ""
authors = [
    {name = "Mara Sarateanu",email = "maraisarateanu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit (>=1.51.0,<2.0.0)",
    "google-auth (>=2.43.0,<3.0.0)",
    "google-cloud-storage (>=3.6.0,<4.0.0)",
    "google-cloud-bigquery (>=3.38.0,<4.0.0)",
    "plotly (>=6.5.0,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "db-dtypes (>=1.4.4,<2.0.0)",
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
